+++ lib/src/widgets/slidable.dart
@@ -107,7 +107,7 @@ class SlidableDelegateContext {
   const SlidableDelegateContext(
     this.slidable,
     this.showActions,
-    this.dragExtent,
+    this.dragSign,
     this.controller,
   );
 
@@ -126,7 +126,7 @@ class SlidableDelegateContext {
   /// Whether the actions have to be shown.
   final bool showActions;
 
-  final double dragExtent;
+  final double dragSign;
 
   /// The animation controller which value depends on  `dragExtent`.
   final AnimationController controller;
@@ -199,10 +199,10 @@ abstract class SlidableStackDelegate extends SlidableDelegate {
   Widget buildActions(BuildContext context, SlidableDelegateContext ctx) {
     final animation = new Tween(
       begin: Offset.zero,
-      end: ctx.createOffset(ctx.totalActionsExtent * ctx.dragExtent.sign),
+      end: ctx.createOffset(ctx.totalActionsExtent * ctx.dragSign),
     ).animate(ctx.controller);
 
-    if (ctx.controller.value != .0 && ctx.dragExtent != .0) {
+    if (ctx.controller.value != .0) {
       return new Container(
         child: new Stack(
           children: <Widget>[
@@ -237,7 +237,7 @@ class SlidableStrechDelegate extends SlidableStackDelegate {
   Widget buildStackActions(BuildContext context, SlidableDelegateContext ctx) {
     final animation = new Tween(
       begin: Offset.zero,
-      end: ctx.createOffset(ctx.totalActionsExtent * ctx.dragExtent.sign),
+      end: ctx.createOffset(ctx.totalActionsExtent * ctx.dragSign),
     ).animate(ctx.controller);
 
     return new Positioned.fill(
@@ -395,12 +395,12 @@ class SlidableDrawerDelegate extends SlidableStackDelegate {
 ///
 /// By sliding in one of these direction, slide actions will appear.
 class Slidable extends StatefulWidget {
-  /// Creates a widget that can be dismissed.
+  /// Creates a widget that can be slid.
   ///
   /// The [actions] contains the slide actions that appears when the child has been dragged down or to the right.
   /// The [secondaryActions] contains the slide actions that appears when the child has been dragged up or to the left.
   ///
-  /// The [delegate] argument must not be null. The [actionExtentRatio]
+  /// The [delegate] and [closeOnScroll] arguments must not be null. The [actionExtentRatio]
   /// and [showAllActionsThreshold] arguments must be greater or equal than 0 and less or equal than 1.
   Slidable({
     Key key,
@@ -412,6 +412,7 @@ class Slidable extends StatefulWidget {
     double actionExtentRatio = _kActionsExtentRatio,
     Duration movementDuration = const Duration(milliseconds: 200),
     Axis direction = Axis.horizontal,
+    bool closeOnScroll = true,
   }) : this.builder(
           key: key,
           child: child,
@@ -423,8 +424,16 @@ class Slidable extends StatefulWidget {
           actionExtentRatio: actionExtentRatio,
           movementDuration: movementDuration,
           direction: direction,
+          closeOnScroll: closeOnScroll,
         );
 
+  /// Creates a widget that can be slid.
+  ///
+  /// The [actionDelegate] is a delegate that builds the slide actions that appears when the child has been dragged down or to the right.
+  /// The [secondaryActionDelegate] is a delegate that builds the slide actions that appears when the child has been dragged up or to the left.
+  ///
+  /// The [delegate] and [closeOnScroll] arguments must not be null. The [actionExtentRatio]
+  /// and [showAllActionsThreshold] arguments must be greater or equal than 0 and less or equal than 1.
   Slidable.builder({
     Key key,
     @required this.child,
@@ -435,6 +444,7 @@ class Slidable extends StatefulWidget {
     this.actionExtentRatio = _kActionsExtentRatio,
     this.movementDuration = const Duration(milliseconds: 200),
     this.direction = Axis.horizontal,
+    this.closeOnScroll = true,
   })  : assert(delegate != null),
         assert(direction != null),
         assert(
@@ -447,6 +457,7 @@ class Slidable extends StatefulWidget {
                 actionExtentRatio >= .0 &&
                 actionExtentRatio <= 1.0,
             'actionExtentRatio must be between 0.0 and 1.0'),
+        assert(closeOnScroll != null),
         super(key: key);
 
   /// The widget below this widget in the tree.
@@ -479,6 +490,16 @@ class Slidable extends StatefulWidget {
   /// Defines the duration for card to go to final position or to come back to original position if threshold not reached.
   final Duration movementDuration;
 
+  /// Specifies to close this slidable after the closest [Scrollable]'s position changed.
+  ///
+  /// Defaults to true.
+  final bool closeOnScroll;
+
+  /// The state from the closest instance of this class that encloses the given context.
+  static SlidableState of(BuildContext context) {
+    return context.ancestorStateOfType(const TypeMatcher<SlidableState>());
+  }
+
   @override
   SlidableState createState() => SlidableState();
 }
@@ -493,24 +514,20 @@ class SlidableState extends State<Slidable>
           ..addStatusListener(_handleShowAllActionsStatusChanged);
   }
 
-  void _handleShowAllActionsStatusChanged(AnimationStatus status) {
-    if (status == AnimationStatus.completed && !_dragUnderway && !_opening) {
-      _dragExtent = .0;
-    }
-
-    if (status == AnimationStatus.completed) {
-      setState(() {});
-    }
-  }
-
   AnimationController _controller;
+
   double _dragExtent = 0.0;
-  bool _dragUnderway = false;
-  bool _opening = false;
+
+  ScrollPosition _scrollPosition;
 
   bool get _showActions {
     return _dragExtent > 0;
   }
+
+  @override
+  bool get wantKeepAlive =>
+      _controller != null &&
+      (_controller.isAnimating || _controller.isCompleted);
 
   /// The current actions that have to be shown.
   SlideActionDelegate get actionDelegate =>
@@ -527,13 +544,61 @@ class SlidableState extends State<Slidable>
   }
 
   @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _removeScrollingNotifierListener();
+    _addScrollingNotifierListener();
+  }
+
+  @override
+  void didUpdateWidget(Slidable oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    if (widget.closeOnScroll != oldWidget.closeOnScroll) {
+      _removeScrollingNotifierListener();
+      _addScrollingNotifierListener();
+    }
+  }
+
+  void _addScrollingNotifierListener() {
+    if (widget.closeOnScroll) {
+      _scrollPosition = Scrollable.of(context)?.position;
+      if (_scrollPosition != null)
+        _scrollPosition.isScrollingNotifier.addListener(_isScrollingListener);
+    }
+  }
+
+  void _removeScrollingNotifierListener() {
+    if (_scrollPosition != null) {
+      _scrollPosition.isScrollingNotifier.removeListener(_isScrollingListener);
+    }
+  }
+
+  @override
   void dispose() {
     _controller.dispose();
+    _removeScrollingNotifierListener();
     super.dispose();
   }
 
+  void open() {
+    _controller.fling(velocity: 1.0);
+  }
+
+  void close() {
+    _controller.fling(velocity: -1.0);
+  }
+
+  void _isScrollingListener() {
+    if (!widget.closeOnScroll || _scrollPosition == null) return;
+
+    // When a scroll starts close this.
+    if (_scrollPosition.isScrollingNotifier.value) {
+      close();
+    }
+  }
+
   void _handleDragStart(DragStartDetails details) {
-    _dragUnderway = true;
     _dragExtent = _controller.value * _overallDragAxisExtent * _dragExtent.sign;
     if (_controller.isAnimating) {
       _controller.stop();
@@ -549,21 +614,23 @@ class SlidableState extends State<Slidable>
   }
 
   void _handleDragEnd(DragEndDetails details) {
-    _dragUnderway = false;
     final double velocity = details.primaryVelocity;
-    final bool open = velocity.sign == _dragExtent.sign;
+    final bool shouldOpen = velocity.sign == _dragExtent.sign;
     final bool fast = velocity.abs() > widget.delegate.fastThreshold;
-    if (!open && fast) {
-      _opening = false;
-      _controller.animateTo(0.0);
-    } else if (_controller.value >= widget.showAllActionsThreshold ||
-        (open && fast)) {
-      _opening = true;
-      _controller.animateTo(1.0);
+    if (_controller.value >= widget.showAllActionsThreshold ||
+        (shouldOpen && fast)) {
+      open();
     } else {
-      _opening = false;
-      _controller.animateTo(0.0);
+      close();
     }
+  }
+
+  void _handleShowAllActionsStatusChanged(AnimationStatus status) {
+    if (status == AnimationStatus.completed) {
+      setState(() {});
+    }
+
+    updateKeepAlive();
   }
 
   @override
@@ -590,7 +657,7 @@ class SlidableState extends State<Slidable>
         new SlidableDelegateContext(
           widget,
           _showActions,
-          _dragExtent,
+          _dragExtent.sign,
           _controller,
         ),
       );
@@ -607,7 +674,4 @@ class SlidableState extends State<Slidable>
       child: content,
     );
   }
-
-  @override
-  bool get wantKeepAlive => _opening;
 }