+++ lib/src/widgets/slide_action.dart
@@ -1,15 +1,69 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter_slidable/src/widgets/slidable.dart';
+
+const bool _kCloseOnTap = true;
+
+/// Abstract class for slide actions that can close after [onTap] occurred.
+abstract class ClosableSlideAction extends StatelessWidget {
+  /// Creates a slide that closes when a tap has occurred if [closeOnTap]
+  /// is [true].
+  ///
+  /// The [closeOnTap] argument must not be null.
+  const ClosableSlideAction({
+    Key key,
+    this.onTap,
+    this.closeOnTap = _kCloseOnTap,
+  })  : assert(closeOnTap != null),
+        super(key: key);
+
+  /// A tap has occurred.
+  final VoidCallback onTap;
+
+  /// Whether close this after tap occurred.
+  ///
+  /// Defaults to true.
+  final bool closeOnTap;
+
+  /// Calls [onTap] if not null and closes the closest [Slidable]
+  /// that encloses the given context.
+  void _handleCloseAfterTap(BuildContext context) {
+    if (onTap != null) {
+      onTap();
+    }
+
+    Slidable.of(context).close();
+  }
+
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: !closeOnTap ? onTap : () => _handleCloseAfterTap(context),
+      child: buildAction(context),
+    );
+  }
+
+  Widget buildAction(BuildContext context);
+}
 
 /// A basic slide action with a background color and a child that will
 /// be center inside its area.
-class SlideAction extends StatelessWidget {
+class SlideAction extends ClosableSlideAction {
+  /// Creates a slide action with a child.
+  ///
+  /// The `color` argument is a shorthand for `decoration: new
+  /// BoxDecoration(color: color)`, which means you cannot supply both a `color`
+  /// and a `decoration` argument. If you want to have both a `color` and a
+  /// `decoration`, you can pass the color as the `color` argument to the
+  /// `BoxDecoration`.
+  ///
+  /// The [closeOnTap] argument must not be null.
   SlideAction({
     Key key,
     @required this.child,
-    this.onTap,
+    VoidCallback onTap,
     Color color,
     Decoration decoration,
+    bool closeOnTap = _kCloseOnTap,
   })  : assert(child != null),
         assert(decoration == null || decoration.debugAssertIsValid()),
         assert(
@@ -18,46 +72,58 @@ class SlideAction extends StatelessWidget {
             'The color argument is just a shorthand for "decoration: new BoxDecoration(color: color)".'),
         decoration = decoration ??
             (color != null ? new BoxDecoration(color: color) : null),
-        super(key: key);
+        super(
+          key: key,
+          onTap: onTap,
+          closeOnTap: closeOnTap,
+        );
 
   final Decoration decoration;
 
   final Widget child;
 
-  final VoidCallback onTap;
-
   @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        decoration: decoration,
-        child: new Center(
-          child: child,
-        ),
+  Widget buildAction(BuildContext context) {
+    return Container(
+      decoration: decoration,
+      child: new Center(
+        child: child,
       ),
     );
   }
 }
 
 /// A basic slide action with an icon, a caption and a background color.
-class IconSlideAction extends StatelessWidget {
+class IconSlideAction extends ClosableSlideAction {
+  /// Creates a slide action with an icon, a [caption] if set and a
+  /// background color.
+  ///
+  /// The [closeOnTap] argument must not be null.
   const IconSlideAction({
     Key key,
     @required this.icon,
     this.caption,
     Color color,
-    this.onTap,
+    VoidCallback onTap,
+    bool closeOnTap = _kCloseOnTap,
   })  : color = color ?? Colors.white,
-        super(key: key);
+        super(
+          key: key,
+          onTap: onTap,
+          closeOnTap: closeOnTap,
+        );
 
   final IconData icon;
+
   final String caption;
+
+  /// The background color.
+  ///
+  /// Defaults to true.
   final Color color;
-  final VoidCallback onTap;
 
   @override
-  Widget build(BuildContext context) {
+  Widget buildAction(BuildContext context) {
     final Color foregroundColor =
         ThemeData.estimateBrightnessForColor(color) == Brightness.light
             ? Colors.black
@@ -71,23 +137,20 @@ class IconSlideAction extends StatelessWidget {
           .caption
           .copyWith(color: foregroundColor),
     );
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        color: color,
-        child: new Center(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              new Flexible(
-                child: new Icon(
-                  icon,
-                  color: foregroundColor,
-                ),
+    return Container(
+      color: color,
+      child: new Center(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            new Flexible(
+              child: new Icon(
+                icon,
+                color: foregroundColor,
               ),
-              new Flexible(child: textWidget),
-            ],
-          ),
+            ),
+            new Flexible(child: textWidget),
+          ],
         ),
       ),
     );